[refactor] 테스트용 메서드 MockupController로 이동 #29
- MockUpController.java : 테스트용 메서드 이동
