+++ BE/todo-12/src/main/java/dev/codesquad/java/todo12/ApiCategoryController.java
@@ -23,8 +23,7 @@ public class ApiCategoryController {
 
     @GetMapping("")
     public ResponseEntity viewAll() {
-        return new ResponseEntity(categoryRepository.findAllDeletedFalse(), HttpStatus.OK);
-        //return new ResponseEntity(getKanban(1L), HttpStatus.OK);
+        return new ResponseEntity(getKanban(1L), HttpStatus.OK);
         //return new ResponseEntity(categoryRepository.findAll(), HttpStatus.OK);
     }
