+++ BE/todo-12/src/main/java/dev/codesquad/java/todo12/MockUpController.java
@@ -3,7 +3,10 @@ package dev.codesquad.java.todo12;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -42,5 +45,45 @@ public class MockUpController {
         kanban.addCategory(category2);
 
         return kanban;
+    }
+
+    @GetMapping("/move/{categoryId}/{id}/{index}")
+    public ResponseEntity move2(@PathVariable Long categoryId, @PathVariable Long id, @PathVariable int index) {
+        Category toCategory = getCategory(categoryId);
+        Card card = getCard(id);
+        cardRepository.delete(card);
+        toCategory.addCard(index, card);
+        categoryRepository.save(toCategory);
+        return new ResponseEntity(card, HttpStatus.OK);
+    }
+
+    @GetMapping("/add/{categoryId}")
+    public ResponseEntity addTest(@PathVariable Long categoryId){
+        Card card = new Card("새거", "택배언제와");
+        Category category = getCategory(categoryId);
+        category.addCard(card);
+        categoryRepository.save(category);
+        return new ResponseEntity(category, HttpStatus.OK);
+    }
+
+    @GetMapping("/delete/{id}")
+    public ResponseEntity deleteTest(@PathVariable Long id) {
+        Card card = getCard(id);
+        card.delete();
+        cardRepository.save(card);
+        return new ResponseEntity(card, HttpStatus.OK);
+    }
+
+    @GetMapping("/")
+    public ResponseEntity test() {
+        return new ResponseEntity(categoryRepository.findAllDeletedFalse(), HttpStatus.OK);
+    }
+
+    private Card getCard(Long id) {
+        return cardRepository.findByIdOnlyDeletedFalse(id).orElseThrow(() -> new DataNotFoundException("해당 카드 없음"));
+    }
+
+    private Category getCategory(Long id) {
+        return categoryRepository.findById(id).orElseThrow(() -> new DataNotFoundException("해당 카테고리 없음"));
     }
 }
